backend/requirements.txt
fastapi
uvicorn[standard]
openai
openai-agents==0.0.16
python-dotenv
rich # ログ出力用に残す場合。APIのみなら不要な場合も
sse-starlette
pydantic-settings

# 認証・データベース関連
pyjwt[crypto]
httpx
supabase
stripe

# SerpAPI・スクレイピング関連
serpapi
google-search-results
beautifulsoup4
requests

# agents SDKが内部で使用する可能性のあるライブラリ
griffe
